Hackathone-Phase-II/phase-4/backend/requirements.txt
fastapi>=0.104.1
sqlmodel>=0.0.16
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
SQLAlchemy>=2.0.23
aiosqlite>=0.19.0
openai>=1.12.0
mcp>=1.0.0